New modern, simple and Flat Design Integrated - could add google ads
The Design and layout have been completely changed t keep up with the
heavily changing simple, responsive and flat design. Makes the webpage
feel a lot more comforting and yet more professional! The changes that
have been made are positive and has had a positove effect on the
websites hits. Revamp for Gallery needs to be underway soon.